main/requirements.txt
Flask==3.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.1.3
regex
tqdm
tabulate==0.9.0
redis==5.0.1
celery==5.3.6
spotipy
Pillow